ai_tech_support_agent/requirements.txt
pypdf
python-docx
python-pptx
unstructured[local-inference]
nltk
sentence-transformers
faiss-cpu
transformers
fastapi
uvicorn[standard]
python-dotenv
accelerate
